added index files to Slider and RangeSlider; added Slider and Range slider to the main index file
