+++ src/components/index.ts
@@ -32,3 +32,5 @@ export * from './Link';
 export * from './FormControl';
 export * from './Info';
 export * from './Status';
+export * from './Slider';
+export * from './RangeSlider';
